better handle -e 'command' if there is a shell expression.
If command contains spaces, $ or other known shell commands, it's
spawned with a "sh -c $COMMAND" instead.

this allows us to start terminology to "watch ps" or things like that.

Also changed the return code on failure to 127, similar to system(3).

NOTE: should we reset some signals? Remember that ecore traps most of them.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk/terminology@72313 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33