options infrastructure refactor.
There is no global config option anymore, we have option instances
that are set to the termio and media for future queries.

The idea is that we can later have different configuration per termio,
like multiple windows or tabs with different font size, background,
etc.

It's just a refactor. It would be better to define if things like
background is up to the whole terminal and then save to a different
config entry... or just save the global and copy it for next termio,
defining the other instances as 'temporary'. Anyway: policy to be defined.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk/terminology@72315 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33